Move Windows API enums to common
