+++ src/api/windows_api.rs
@@ -1,45 +1,8 @@
-use crate::common::{MonitorHandle, MonitorInfo, Monitors, Point, Rect, Window, WindowHandle, WindowPlacement};
-use std::fmt::{Display, Formatter};
+use crate::common::{
+  MonitorHandle, MonitorInfo, Monitors, Point, Rect, Window, WindowHandle, WindowLookupError, WindowMetadata,
+  WindowPlacement, WindowPositioningResult,
+};
 use windows::Win32::UI::Shell::IVirtualDesktopManager;
-
-/// Information about a visible top-level window.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct WindowMetadata {
-  pub handle: WindowHandle,
-  pub title: String,
-  pub class_name: String,
-  pub rect: Rect,
-}
-
-/// Result of atomically positioning a window batch.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum WindowPositioningResult {
-  Applied,
-  Rejected(Vec<WindowHandle>),
-  BatchFailed,
-}
-
-/// Failure to identify a visible top-level window.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum WindowLookupError {
-  NoTarget,
-  Vanished,
-  OwnWindow,
-  AccessDenied,
-}
-
-impl Display for WindowLookupError {
-  fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::NoTarget => write!(f, "No window was found at that point"),
-      Self::Vanished => write!(f, "The selected window is no longer available"),
-      Self::OwnWindow => write!(f, "Randolf's Window Picker cannot select itself"),
-      Self::AccessDenied => write!(f, "Randolf does not have permission to inspect that window"),
-    }
-  }
-}
-
-impl std::error::Error for WindowLookupError {}
 
 pub trait WindowsApi {
   fn is_running_as_admin(&self) -> bool;
