+++ src/common/window_positioning_result.rs
@@ -1,0 +1,9 @@
+use crate::common::WindowHandle;
+
+/// Result of atomically positioning a window batch.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum WindowPositioningResult {
+  Applied,
+  Rejected(Vec<WindowHandle>),
+  BatchFailed,
+}
